test: split #2191 chrome tests out of the oversized screen suite (#2191)
Keep TmuxSessionScreenTest.kt under the 128KiB ratchet. New
TmuxSessionScreenChromeTest holds the four chrome-reachability tests.
